Mudando para double

+++ BeHappy/BeHappy/Models/IOperacao.cs
@@ -7,9 +7,9 @@ namespace BeHappy.Models
 {
     public interface IOperacao //Service
     {
-        public int soma();
-        public int subracao();
-        public int multiplicacao();
-        public int divisao();
+        public double soma();
+        public double subracao();
+        public double multiplicacao();
+        public double divisao();
     }
 }
